refactor(frontend): Add separate components for schema, election info and positions
