+++ frontend/src/components/ElectionSetup/ElectionSetup.tsx
@@ -1,19 +1,13 @@
-import { Button, Input, DatePicker, Checkbox } from '@heroui/react';
+import { Button } from '@heroui/react';
 import { now, getLocalTimeZone } from '@internationalized/date';
 import type { ZonedDateTime } from '@internationalized/date';
 import { useState } from 'react';
 import useSWRMutation from 'swr/mutation';
-import { z } from 'zod';
 
-import { fetcher } from '../lib/fetcher';
-
-export const electionSchema = z.object({
-	name: z.string().min(1),
-	nomination_start: z.date(),
-	nomination_end: z.date(),
-	voting_start: z.date(),
-	voting_end: z.date(),
-});
+import { fetcher } from '../../lib/fetcher';
+import ElectionInfo from './ElectionInfo';
+import Positions from './Positions';
+import { electionSchema } from './schema';
 
 export const ElectionSetup = () => {
 	const [electionName, setElectionName] = useState('');
@@ -162,128 +156,26 @@ export const ElectionSetup = () => {
 				<div className="flex-grow border-t border-gray-400"></div>
 			</div>
 			<h2 className="mb-4 text-lg font-semibold">Election Info</h2>
-			<div className="flex flex-col gap-4">
-				<div className="flex flex-col gap-2">
-					<Input
-						label="Election Name"
-						placeholder="Enter election name"
-						type="text"
-						value={electionName}
-						onChange={(e) => setElectionName(e.target.value)}
-						errorMessage="Election name is required"
-						isInvalid={!!errors.electionName}
-					/>
-				</div>
-				<div className="flex w-full items-end gap-3">
-					<DatePicker
-						hideTimeZone
-						showMonthAndYearPickers
-						defaultValue={nominationStartDate}
-						label="Nomination Start Date"
-						onChange={(date) => date && setNominationStartDate(date)}
-						errorMessage={errors.nominationStartDate}
-						isInvalid={!!errors.nominationStartDate}
-					/>
-					<DatePicker
-						hideTimeZone
-						showMonthAndYearPickers
-						defaultValue={nominationEndDate}
-						label="Nomination End Date"
-						onChange={(date) => date && setNominationEndDate(date)}
-						errorMessage={errors.nominationEndDate}
-						isInvalid={!!errors.nominationEndDate}
-					/>
-				</div>
-				<div className="flex w-full items-end gap-3">
-					<DatePicker
-						hideTimeZone
-						showMonthAndYearPickers
-						defaultValue={votingStartDate}
-						label="Voting Start Date"
-						onChange={(date) => date && setVotingStartDate(date)}
-						errorMessage={errors.votingStartDate}
-						isInvalid={!!errors.votingStartDate}
-					/>
-					<DatePicker
-						hideTimeZone
-						showMonthAndYearPickers
-						defaultValue={votingEndDate}
-						label="Voting End Date"
-						onChange={(date) => date && setVotingEndDate(date)}
-						errorMessage={errors.votingEndDate}
-						isInvalid={!!errors.votingEndDate}
-					/>
-				</div>
-			</div>
+			<ElectionInfo
+				electionName={electionName}
+				setElectionName={setElectionName}
+				errors={errors}
+				nominationStartDate={nominationStartDate}
+				setNominationStartDate={setNominationStartDate}
+				nominationEndDate={nominationEndDate}
+				setNominationEndDate={setNominationEndDate}
+				votingStartDate={votingStartDate}
+				setVotingStartDate={setVotingStartDate}
+				votingEndDate={votingEndDate}
+				setVotingEndDate={setVotingEndDate}
+			/>
 			<div className="h-4"></div>
-			<div className="mb-4 flex items-center justify-between">
-				<h2 className="text-lg font-semibold">Positions</h2>
-				<Button color="primary" onPress={addPosition}>
-					<span className="mr-0.5 text-2xl">+</span> Add Position
-				</Button>
-			</div>
-			{positions.map((position, index) => (
-				<div
-					key={index}
-					className="flex flex-col gap-4 rounded-xl bg-gray-200 p-4"
-				>
-					<Input
-						label="Position Name"
-						placeholder="Enter position name"
-						type="text"
-						value={position.name}
-						onChange={(e) =>
-							updatePosition(index, { ...position, name: e.target.value })
-						}
-					/>
-					<Input
-						label="Description"
-						placeholder="Enter position description"
-						type="text"
-						value={position.description}
-						onChange={(e) =>
-							updatePosition(index, {
-								...position,
-								description: e.target.value,
-							})
-						}
-					/>
-					<div className="flex items-center gap-2 md:gap-6">
-						<Input
-							label="Vacancies"
-							placeholder="Enter number of vacancies"
-							type="number"
-							value={position.vacancies.toString()}
-							min={1}
-							onChange={(e) =>
-								updatePosition(index, {
-									...position,
-									vacancies: parseInt(e.target.value),
-								})
-							}
-							className="w-1/3 md:w-full"
-						/>
-						<Checkbox
-							checked={position.executive}
-							onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-								updatePosition(index, {
-									...position,
-									executive: e.target.checked,
-								})
-							}
-						>
-							Executive
-						</Checkbox>
-						<Button
-							color="secondary"
-							onPress={() => removePosition(index)}
-							className="w-[40%] md:w-1/6"
-						>
-							<span className="mr-0.5 text-2xl">-</span> Remove Position
-						</Button>
-					</div>
-				</div>
-			))}
+			<Positions
+				addPosition={addPosition}
+				positions={positions}
+				updatePosition={updatePosition}
+				removePosition={removePosition}
+			/>
 			<div className="flex justify-center">
 				<Button color="primary" className="mt-4" onPress={handleSubmit}>
 					Create Election
